INS-1644 moving&refactoring
