+++ ledger/storage/jet/drop/dropbuilder.go
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package jet
+package drop
 
 import (
 	"context"
@@ -23,6 +23,7 @@ import (
 	"github.com/dgraph-io/badger"
 	"github.com/insolar/insolar/core"
 	"github.com/insolar/insolar/ledger/storage"
+	"github.com/insolar/insolar/ledger/storage/jet"
 	"github.com/insolar/insolar/ledger/storage/record"
 	"github.com/pkg/errors"
 )
@@ -34,7 +35,7 @@ type Builder interface {
 	PrevHash(prevHash []byte)
 	Pulse(pn core.PulseNumber)
 
-	Build() (JetDrop, error)
+	Build() (jet.JetDrop, error)
 }
 
 // Hashable is a base interface for an item, that can be appended to builder
@@ -78,18 +79,18 @@ func (b *builder) Pulse(pn core.PulseNumber) {
 }
 
 // Build builds JetDrop and returns it
-func (b *builder) Build() (JetDrop, error) {
+func (b *builder) Build() (jet.JetDrop, error) {
 	if b.pn == nil {
-		return JetDrop{}, errors.New("pulseNumber is required")
+		return jet.JetDrop{}, errors.New("pulseNumber is required")
 	}
 	if b.dropSize == nil {
-		return JetDrop{}, errors.New("dropSize is required")
+		return jet.JetDrop{}, errors.New("dropSize is required")
 	}
 	if b.prevHash == nil && *b.pn != core.FirstPulseNumber {
-		return JetDrop{}, errors.New("prevHash is required")
+		return jet.JetDrop{}, errors.New("prevHash is required")
 	}
 
-	return JetDrop{
+	return jet.JetDrop{
 		Pulse:    *b.pn,
 		PrevHash: b.prevHash,
 		Hash:     b.Hasher.Sum(nil),
@@ -100,7 +101,7 @@ func (b *builder) Build() (JetDrop, error) {
 // Packer is an wrapper interface around process of building jetdrop
 // It's considered that implementation of packer uses Bulder under the hood
 type Packer interface {
-	Pack(ctx context.Context, jetID storage.JetID, pulse core.PulseNumber, prevHash []byte) (JetDrop, error)
+	Pack(ctx context.Context, jetID storage.JetID, pulse core.PulseNumber, prevHash []byte) (jet.JetDrop, error)
 }
 
 // NewPacker creates db-based impl of packer
@@ -117,7 +118,7 @@ type packer struct {
 }
 
 // Pack creates new JetDrop through interactions with db and Builder
-func (p *packer) Pack(ctx context.Context, jetID storage.JetID, pulse core.PulseNumber, prevHash []byte) (JetDrop, error) {
+func (p *packer) Pack(ctx context.Context, jetID storage.JetID, pulse core.PulseNumber, prevHash []byte) (jet.JetDrop, error) {
 	p.DBContext.WaitingFlight()
 	_, jetPrefix := jetID.Jet()
 
@@ -143,7 +144,7 @@ func (p *packer) Pack(ctx context.Context, jetID storage.JetID, pulse core.Pulse
 		return nil
 	})
 	if err != nil {
-		return JetDrop{}, err
+		return jet.JetDrop{}, err
 	}
 
 	p.Pulse(pulse)
